refactor(renderer): keep the index to the renderer and publish the standalone pieces on leaves
