+++ src/renderer/index.ts
@@ -1,32 +1,28 @@
-// index.ts is the renderer module's API. `constants.ts`, `themeSettings.ts` and `util.ts` are
-// published alongside it: they import nothing, so a consumer that needs only a class name or a pure
-// helper can take one without pulling the engine into its bundle.
+// The index publishes the renderer, the leaves publish the standalone pieces. `constants.ts`,
+// `text.ts`, `themeSettings.ts` and `util.ts` are the leaves: they import nothing, so a consumer
+// that needs only a class name or a pure helper can take one without pulling the engine into its
+// bundle.
 //
-// `createLyricsRenderer` is the way in. Everything beside it is something one instance cannot answer
-// for on its own: the song level operations address every live view at once, and the injection and
-// text helpers decorate lines that are already built.
+// `createLyricsRenderer` is the way in. The four values beside it are what one instance cannot
+// answer for on its own: the song level operations address every live view at once, and the
+// injection helpers decorate lines that are already built.
 
 export { resetPlaybackClock, resumeAllAutoscroll } from "./engine";
-export { hasNonLatinLyrics, injectRomanization, injectTranslation, type LineData, type PartData } from "./inject";
+export { injectRomanization, injectTranslation, type LineData } from "./inject";
 export { createLyricsRenderer } from "./renderer";
-export { containsNonLatin, detectNonLatinLanguage } from "./text";
-export type { Lyric, LyricsRenderer, LyricsRendererHost, LyricSyncType, TickOptions } from "./types";
+export type { Lyric, LyricsRenderer, LyricsRendererHost, TickOptions } from "./types";
 
 // -- Published without a consumer here --------------------------
 //
 // `@public` keeps knip off an export this repo does not itself take from this file, so that it goes
-// on reporting the rest. Two kinds qualify, and nothing else should: a symbol the extension imports
-// from its own published leaf instead, which a package consumer would still expect to find on the
-// index, and a type named in the signature of something published, which a consumer has to be able
-// to spell.
+// on reporting the rest. One kind qualifies, and nothing else should: a type named in the shape of
+// something published, which a consumer has to be able to spell.
 
 /** @public */
 export type { AnimationTickStatus } from "./engine";
 /** @public */
-export { registerThemeSetting, setThemeSettings } from "./themeSettings";
+export type { PartData } from "./inject";
 /** @public */
-export type { LyricPart, LyricsRendererDebugSink, LyricsRendererOptions } from "./types";
-/** @public */
-export { reflow, toMs } from "./util";
+export type { LyricPart, LyricsRendererDebugSink, LyricsRendererOptions, LyricSyncType } from "./types";
 /** @public */
 export type { SetLyricsOptions } from "./view";
